requirements.txt
praw
python-dotenv
openai
wordcloud
matplotlib
